+++ pol-core/clib/rawtypes.h
@@ -26,11 +26,11 @@ Notes
 
 typedef unsigned char     u8;
 typedef unsigned short    u16;
-typedef unsigned long     u32;
+typedef unsigned int      u32;
 
 typedef signed char       s8;
 typedef signed short      s16;
-typedef signed long       s32;
+typedef signed int        s32;
 
 #ifdef _WIN32
 typedef unsigned __int64    u64;
